requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AI and LangChain
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
azure-identity==1.15.0
openai>=1.6.1,<2.0.0

# MCP Protocol
mcp>=1.12.4,<2.0.0  # Official MCP Python SDK (FastMCP compatible)

# HTTP and API
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
jinja2==3.1.2
click==8.1.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0

# File Processing
python-magic==0.4.27
chardet==5.2.0

# Web Scraping
beautifulsoup4==4.12.2
lxml==4.9.3

# Data Processing
pandas==2.1.4
numpy==1.25.2
